Document regenerating HTML with nodemon
I found this useful so I thought it was worth sharing.  In addition, because `spec-md` lacks a flag
to specify an output file (and instead depends on redirecting to a file), I found it necessary/convenient
to leverage NPM scripts to avoid the need for redirection in the `nodemon` invocation (which would have
sent `nodemon` output to a file, not the output from `spec-md`).