Add content to `<head>`
This allows a new `head` option (via node API or within metadata.json) which provides the ability to insert any arbitrary string within the `<head>` tag of the output HTML.

This offers a way to add additional styles, scripts, and meta tags. As an example, you might use this feature to introduce KaTeX or MathJax.

Fixes #15