Subsections!
Subsections are a new addition to spec-md which captures a common pattern with a significant usage: a one-line paragraph of bold text as a header for a section of content which is not significant enough to warrant being in the table of contents.

This pattern already appeared in the spec-md documentation, and appears frequently in both ECMAScript and GraphQL specifications.

As part of introducing subsections, this cleans up the typography for section headers.