Support standard markdown titles (#44)
Supports use of standard markdown headers (`# Header`) as the title of a spec markdown document. Inspired by a similar change in #31 however this does not require wrapping the header in bold.

For some context, originally setext headers were required to differentiate between a document title and a section header in an imported file since spec-md did not differentiate between initial files and imported files. This change is possible by having pegjs support multiple starting positions for the parser and parse initial and imported documents differently.

This also allows us to remove the ast validation logic since that can now be encoded directly in the grammar.