Resolve link urls from biblio
If a markdown link has a hash reference that resolves in the biblio, then link to where biblio wants to link. Also look for `sec-` and `()` variants to properly convert from links which would resolve when viewed in Github to links that would resolve within a printed spec md html file.

Fixes #32