githubSource links
This adds an option to provide a github base url and then uses source map data to link to the original files.

This is especially helpful for drafts or other editions of spec documents where the workflow to go from rendered result back to original source in repo is common.