Block based grammar
This is an iteration of the PEG grammar which seeks to start each block rule on the line in which it begins (rather than at the end of the previously matched rule). A new top level set of "blocks" rules are added which skip over new lines and empty lines.